pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "playlistpipe"
version = "0.1.0"
description = "Turn YouTube playlists into Notion databases, Obsidian notes, or Anki decks."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"   # tomllib is stdlib from 3.11
authors = [{ name = "Srinath Murali" }]
keywords = ["youtube", "notion", "obsidian", "anki", "playlist", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Utilities",
]
dependencies = [
    "requests>=2.31",
    "genanki>=0.13",
    "yt-dlp>=2024.08.06",
    "questionary>=2.0",
    "tomli-w>=1.0",
]

[project.optional-dependencies]
clipboard = ["pyperclip>=1.8"]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-requests",
    "pyyaml",   # for obsidian tests
]

[project.scripts]
plp = "playlistpipe.cli:main"
playlistpipe = "playlistpipe.cli:main"

[project.urls]
Homepage = "https://github.com/srimur/playlistpipe"
Issues = "https://github.com/srimur/playlistpipe/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/playlistpipe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "PTH", "RUF"]
ignore = ["E501"]   # line length handled by formatter

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
disallow_untyped_defs = true
packages = ["playlistpipe"]
